Moved: InputSubstitution to Calamari.Util
